+++ app/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
@@ -16,21 +17,33 @@ class ProductController extends Controller
      */
     public function product()
     {
-        $products = Product::paginate(5);
-        return view('admin.product.products', [
-            'products' => $products
-        ]);
+        $usertype = Auth::user()->usertype;
+        if ($usertype == '1') {
+
+            $products = Product::all();
+            return view('admin.product.products', [
+                'products' => $products
+            ]);
+        } else {
+            return redirect()->route('login');
+        }
     }
 
 
     public function addproduct()
     {
-        $categories = Category::all();
-        $products = Product::all();
-        return view('admin.product.addproduct', [
-            'products' => $products,
-            'categories' => $categories
-        ]);
+        $usertype = Auth::user()->usertype;
+        if ($usertype == '1') {
+
+            $categories = Category::all();
+            $products = Product::all();
+            return view('admin.product.addproduct', [
+                'products' => $products,
+                'categories' => $categories
+            ]);
+        } else {
+            return redirect()->route('login');
+        }
     }
     /**
      * Show the form for creating a new resource.
@@ -50,52 +63,32 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'product_name' => 'required',
-            'product_price' => 'required',
-            'product_category' => 'required',
+        $usertype = Auth::user()->usertype;
+        if ($usertype == '1') {
+
+            $this->validate($request, [
+                'product_name' => 'required',
+                'product_price' => 'required',
+                'product_category' => 'required',
 
 
-        ]);
-        $product = new product();
-        $imageName = time() . '.' . $request->image->extension();
-        $request->image->move(public_path('product_poster'), $imageName);;
-        $product->poster_url = $imageName;
+            ]);
+            $product = new product();
+            $imageName = time() . '.' . $request->image->extension();
+            $request->image->move(public_path('product_poster'), $imageName);;
+            $product->poster_url = $imageName;
 
-        $product->name = $request->input('product_name');
-        $product->price = $request->input('product_price');
-        $product->description = $request->input('product_description');
-        $product->category_id = $request->input('product_category');
-        $product->status = 1;
-        $product->save();
-        return back()->with('status', 'product created successful');
-
-        // if ($request->hasFile('product_image')) {
-
-        //     $fileNameExt = $request->file('product_image')->getClientOriginalName();
-
-        //     $fileName = pathinfo($fileNameExt, PATHINFO_FILENAME);
-
-        //     $ext = $request->file('product_image')->getClientOriginalExtension();
-
-        //     $fileNameToStore = $fileName . '_' . time() . '.' . $ext;
-
-        //     $path = $request->file('product_image')->storeAs('public/product_images', $fileNameToStore);
-        // } else {
-        //     $fileNameToStore = 'default.png';
-        // }
-
-        // $product = new product();
-        // $product->name = $request->input('product_name');
-        // $product->price = $request->input('product_price');
-        // $product->description = $request->input('product_description');
-        // $product->category_id = $request->input('product_category');
-        // $product->poster_url = $fileNameToStore;
-        // $product->status = 1;
-        // $product->save();
-        // return back()->with('status', 'product created successful');
+            $product->name = $request->input('product_name');
+            $product->price = $request->input('product_price');
+            $product->description = $request->input('product_description');
+            $product->category_id = $request->input('product_category');
+            $product->status = 1;
+            $product->save();
+            return back()->with('status', 'product created successful');
+        } else {
+            return redirect()->route('login');
+        }
     }
-
     /**
      * Display the specified resource.
      *
@@ -115,14 +108,19 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
-        $product = Product::find($id);
+        $usertype = Auth::user()->usertype;
+        if ($usertype == '1') {
+            $product = Product::find($id);
 
-        $categories = Category::all();
+            $categories = Category::all();
 
-        return view('admin.product.editproduct', [
-            'product' => $product,
-            'categories' => $categories
-        ]);
+            return view('admin.product.editproduct', [
+                'product' => $product,
+                'categories' => $categories
+            ]);
+        } else {
+            return redirect()->route('login');
+        }
     }
 
     /**
@@ -134,42 +132,47 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'product_name' => 'required',
-            'product_price' => 'required',
-            'product_description' => 'required',
-            'product_category' => 'required',
-        ]);
+        $usertype = Auth::user()->usertype;
+        if ($usertype == '1') {
+            $this->validate($request, [
+                'product_name' => 'required',
+                'product_price' => 'required',
+                'product_description' => 'required',
+                'product_category' => 'required',
+            ]);
 
-        $product = Product::find($id);
+            $product = Product::find($id);
 
-        $product->name = $request->input('product_name');
-        $product->price = $request->input('product_price');
-        $product->price = $request->input('product_description');
-        $product->category_id = $request->input('product_category');
+            $product->name = $request->input('product_name');
+            $product->price = $request->input('product_price');
+            $product->price = $request->input('product_description');
+            $product->category_id = $request->input('product_category');
 
-        if ($request->hasFile('product_image')) {
-            $fileNameExt = $request->file('product_image')->getClientOriginalName();
+            if ($request->hasFile('product_image')) {
+                $fileNameExt = $request->file('product_image')->getClientOriginalName();
 
-            $fileName = pathinfo($fileNameExt, PATHINFO_FILENAME);
+                $fileName = pathinfo($fileNameExt, PATHINFO_FILENAME);
 
-            $ext = $request->file('product_image')->getClientOriginalExtension();
+                $ext = $request->file('product_image')->getClientOriginalExtension();
 
-            $fileNameToStore = $fileName . '_' . time() . '.' . $ext;
+                $fileNameToStore = $fileName . '_' . time() . '.' . $ext;
 
-            $path = $request->file('product_image')->storeAs('public/product_images', $fileNameToStore);
+                $path = $request->file('product_image')->storeAs('public/product_images', $fileNameToStore);
 
-            //supprimer l'ancienne photo si c'est pas le default.png
-            // pck le default.pgn doit être permanent
-            if ($product->poster_url != 'default.png') {
-                Storage::delete('public/product_images' . $product->poster_url);
+                //supprimer l'ancienne photo si c'est pas le default.png
+                // pck le default.pgn doit être permanent
+                if ($product->poster_url != 'default.png') {
+                    Storage::delete('public/product_images' . $product->poster_url);
+                }
+
+                $product->poster_url = $fileNameToStore;
             }
+            $product->update();
 
-            $product->poster_url = $fileNameToStore;
+            return redirect('/products')->with('status', 'product has been  update successful');
+        } else {
+            return redirect()->route('login');
         }
-        $product->update();
-
-        return redirect('/products')->with('status', 'product has been  update successful');
     }
 
     /**
@@ -180,16 +183,20 @@ class ProductController extends Controller
      */
     public function delete($id)
     {
+        $usertype = Auth::user()->usertype;
+        if ($usertype == '1') {
+            $product = Product::find($id);
 
-        $product = Product::find($id);
+            if ($product->poster_url != 'default.png') {
+                Storage::delete('public/product_images' . $product->poster_url);
+            }
 
-        if ($product->poster_url != 'default.png') {
-            Storage::delete('public/product_images' . $product->poster_url);
+            $product->delete();
+
+            return back()->with('status', 'product has been delete success');
+        } else {
+            return redirect()->route('login');
         }
-
-        $product->delete();
-
-        return back()->with('status', 'product has been delete success');
     }
 
     /**
@@ -200,13 +207,18 @@ class ProductController extends Controller
      */
     public function activer($id)
     {
-        $product = Product::find($id);
+        $usertype = Auth::user()->usertype;
+        if ($usertype == '1') {
+            $product = Product::find($id);
 
-        $product->status = 1;
+            $product->status = 1;
 
-        $product->update();
+            $product->update();
 
-        return back();
+            return back();
+        } else {
+            return redirect()->route('login');
+        }
     }
 
 
@@ -218,12 +230,17 @@ class ProductController extends Controller
      */
     public function desactiver($id)
     {
-        $product = Product::find($id);
+        $usertype = Auth::user()->usertype;
+        if ($usertype == '1') {
+            $product = Product::find($id);
 
-        $product->status = 0;
-        $product->update();
+            $product->status = 0;
+            $product->update();
 
-        return back();
+            return back();
+        } else {
+            return redirect()->route('login');
+        }
     }
 
     public function select_par_category($id)
